# feat: preloading JavaScript files