close #1187

## points

- use `<link rel="modulepreload">` tag and `Link: rel="modulepreload"` header to preload dependencies
- No preloading at all for pages that don't use JavaScript at all.
- Information about preloading is delivered only when the JS file has already been built. If the JS file hasn't finished building, it will respond without preload information instead of waiting for the build and then returning the HTML.

## performance comparison

I've deployed this branch to https://fresh-module-preload.deno.dev/ so you can try it out and see how it behaves.

When I accessed [fresh.deno.dev](https://fresh.deno.dev/) and [fresh-module-preload.deno.dev](https://fresh-module-preload.deno.dev/) 20 times each and measured the DOMContentLoaded time, it was 24% faster (157.3ms -> 118.95ms) in my environment (Nagano City, Japan).

<details>
<summary>Actual measured data</summary>

(Don't forget to check `Disable Cache` in devtools)

 | afeter (ms) | before (ms)
-- | -- | --
0 | 180 | 152
1 | 82 | 144
2 | 78 | 186
3 | 156 | 196
4 | 88 | 211
5 | 85 | 123
6 | 162 | 209
7 | 154 | 110
8 | 86 | 175
9 | 135 | 109
10 | 183 | 117
11 | 115 | 159
12 | 99 | 173
13 | 83 | 131
14 | 82 | 142
15 | 135 | 160
16 | 120 | 125
17 | 148 | 152
18 | 118 | 229
19 | 90 | 143
average | 118.95 | 157.3

</details>

## screen shots

![image](https://github.com/denoland/fresh/assets/40050810/c8f6d46b-77da-4e20-b29c-3b0219907872)

![image](https://github.com/denoland/fresh/assets/40050810/3159904c-07f0-45ee-a9ca-7bed35a292ca)