@ayame113 2023-05-27
Thank you for your review!

> This is great - why both in the headers and the body though? I'd prefer sending it in headers only.

This is for backwards compatibility as the Link header is a new feature.

- The share of browsers that support the Link header is 73%.
    https://caniuse.com/mdn-http_headers_link
    ![image](https://github.com/denoland/fresh/assets/40050810/73f12db9-79e1-428f-902a-58b6f7825799)
- The share of browsers that support `<link rel="modulepreload">` is 96%.
    https://caniuse.com/link-rel-preload
    ![image](https://github.com/denoland/fresh/assets/40050810/338bce3f-5106-4b05-8ff6-13aefd6b6922)

By putting preload information in both the header and body, people whose browsers don't support Link headers can also benefit from preloading.