@iuioiua 2023-05-31
I think we should do one or the other. I prefer using the `<link>` element because it has greater browser support. Outside of that, both perform the same function.