@lucacasonato 2023-05-31
I have rebased the patch and have removed the `Link` headers in favour of just using the meta preload tags.

Can you add some tests?