@ayame113 2023-05-31
OK, I added a test.