@ayame113 2023-05-31
Note: Using the Link header is faster than using the link tag. (For headers, preload can start when response headers arrive. When using tags, response headers arrive -> response body arrives -> parse body -> start preload.)
That said, it's something that can be added later. Merging as is is fine for me!